+++ src/buildTranspilerFile.js
@@ -18,7 +18,6 @@ async function each(settings, name, modules) {
 		base,
 		buildDevDir,
 		buildDistDir,
-		buildTempDir,
 		buildAssetsDir,
 		cssTranspilerAdapter,
 		isDev,
@@ -29,17 +28,21 @@ async function each(settings, name, modules) {
 	let buildDir = isDev ? buildDevDir : buildDistDir;
 	let buildCssFile = `${name}.css`;
 	let entryFile = `${buildCssFile}.${types[0]}`;
+
+	let buildPath = path.join(base, buildDir);
+	await fs.promises.mkdir(buildPath, {recursive: true});
+	let tempPath = await fs.promises.mkdtemp(path.join(buildPath, 'temp-'));
 	
-	let outputPath = path.resolve(path.join(base, buildDir, buildCssFile));
-	let mirrorDir = path.resolve(path.join(base, buildDir, buildAssetsDir));
-	let entryPath = path.resolve(path.join(base, buildDir, buildTempDir, entryFile));
-	let tempMirrorDir = path.resolve(path.join(base, buildDir, buildTempDir, buildAssetsDir));
+	let outputPath = path.resolve(path.join(buildPath, buildCssFile));
+	let mirrorDir = path.resolve(path.join(buildPath, buildAssetsDir));
+	let entryPath = path.resolve(path.join(tempPath, entryFile));
+	let tempMirrorDir = path.resolve(path.join(tempPath, buildAssetsDir));
 	let moduleFiles = await getModuleFilesOfType(base, modules, types);
 	let assetList = [];
 	
 	for (let file of moduleFiles) {
 		let moduleFilePathOriginal = path.join(base, file);
-		let moduleFilePathMirror = path.join(base, buildDir, buildTempDir, buildAssetsDir, file);
+		let moduleFilePathMirror = path.join(tempPath, buildAssetsDir, file);
 		let moduleCode = await relativizeCssUrls(
 			await fs.promises.readFile(moduleFilePathOriginal, 'utf8'),
 			moduleFilePathOriginal,
@@ -61,8 +64,7 @@ async function each(settings, name, modules) {
 	await transpilerAdapter.run({sourcePaths, entryPath, outputPath, isDev});
 	
 	if (!isDev) {
-		let tempDirToDelete = path.resolve(path.join(base, buildDir, buildTempDir));
-		await fs.promises.rm(tempDirToDelete, { recursive: true, force: true });
+		await fs.promises.rm(tempPath, { recursive: true, force: true });
 	}
 	
 	return assetList;