+++ test/test-buildTranspilerFile.js
@@ -8,6 +8,20 @@ const { name, cloneRun } = new DirectoryTester(__filename);
 const expandOptions = require('../src/expandOptions.js');
 const buildTranspilerFile = require('../src/buildTranspilerFile.js');
 
+
+function getTemp(after) {
+	let temp;
+	Object.keys(after).some(item => {
+		let match = item.match(/\/(temp-\w+)\//);
+		if (match) {
+			temp = match[1];
+			return true;
+		}
+		return false;
+	});
+	return temp;
+}
+
 describe(name, async () => {
 
 	it('should send paths to adapter function and return asset paths', async () => {
@@ -23,6 +37,7 @@ describe(name, async () => {
 			let originalAssetList = await buildTranspilerFile(settings);
 
 			let after = box.snapshot();
+			let temp = getTemp(after);
 			let {created} = box.diff(before, after);
 
 			assert.deepEqual(originalAssetList, [
@@ -30,21 +45,21 @@ describe(name, async () => {
 			]);
 
 			assert.deepEqual(created, [
-				'dev/site.css',
-				'dev/temp/assets/src/a/a.css',
-				'dev/temp/assets/src/b/b.scss',
-				'dev/temp/assets/src/c/c.scss',
-				'dev/temp/site.css.scss',
+				`dev/site.css`,
+				`dev/${temp}/assets/src/a/a.css`,
+				`dev/${temp}/assets/src/b/b.scss`,
+				`dev/${temp}/assets/src/c/c.scss`,
+				`dev/${temp}/site.css.scss`,
 			]);
 		
-			assert.equal(after['dev/temp/assets/src/a/a.css'], '#a-css{}');
-			assert.equal(after['dev/temp/assets/src/b/b.scss'], '#b-sass{}');
-			assert.equal(after['dev/temp/assets/src/c/c.scss'],
+			assert.equal(after[`dev/${temp}/assets/src/a/a.css`], '#a-css{}');
+			assert.equal(after[`dev/${temp}/assets/src/b/b.scss`], '#b-sass{}');
+			assert.equal(after[`dev/${temp}/assets/src/c/c.scss`],
 				// relative to {base}/dev, since that's where output.css will be.
 				'#c-sass{ background: url(assets/src/c/icon.gif?h=gJI5Yp!Ng9C6F7mGWXybWDBcL38~) }'
 			);
 		
-			assert.deepEqual(JSON.parse(after['dev/temp/site.css.scss']), {
+			assert.deepEqual(JSON.parse(after[`dev/${temp}/site.css.scss`]), {
 				isEntry: true,
 				isDev: true,
 				sourcePaths: [
@@ -103,13 +118,15 @@ describe(name, async () => {
 			await buildTranspilerFile(settings);
 
 			let after = box.snapshot();
+			
+			let temp = getTemp(after);
 
-			assert.equal(after['dev-weird/temp/zote/sbor/src/c/c.scss'],
+			assert.equal(after[`dev-weird/${temp}/zote/sbor/src/c/c.scss`],
 				// relative to {base}/dev-weird/foo, since that's where output.css will be.
 				'#c-sass{ background: url(../zote/sbor/src/c/icon.gif?h=gJI5Yp!Ng9C6F7mGWXybWDBcL38~) }'
 			);
 		
-			assert.deepEqual(JSON.parse(after['dev-weird/temp/foo/output.css.scss']), {
+			assert.deepEqual(JSON.parse(after[`dev-weird/${temp}/foo/output.css.scss`]), {
 				isEntry: true,
 				isDev: true,
 				sourcePaths: [
@@ -140,13 +157,14 @@ describe(name, async () => {
 			await buildTranspilerFile(settings);
 
 			let after = box.snapshot();
+			let temp = getTemp(after);
 
-			assert.equal(after['dev-weird/temp/zote/sbor/src/c/c.scss'],
+			assert.equal(after[`dev-weird/${temp}/zote/sbor/src/c/c.scss`],
 				// relative to {base}/dev-weird/foo, since that's where output.css will be.
 				'#c-sass{ background: url(../zote/sbor/src/c/icon.gif?h=gJI5Yp!Ng9C6F7mGWXybWDBcL38~) }'
 			);
 		
-			assert.deepEqual(JSON.parse(after['dev-weird/temp/foo/output.css.scss']), {
+			assert.deepEqual(JSON.parse(after[`dev-weird/${temp}/foo/output.css.scss`]), {
 				isEntry: true,
 				isDev: true,
 				sourcePaths: [
@@ -161,6 +179,37 @@ describe(name, async () => {
 			});
 		});
 
+	});
+
+	it('should not collide temp deletions when a collation has no output', async () => {
+		await cloneRun(async (base, box) => { // eslint-disable-line no-unused-vars
+
+			let settings = await expandOptions({
+				collations: [
+					{
+						name: 'one',
+						modules: [ 'src2/a' ],
+					},
+					{
+						name: 'two',
+						modules: [
+							'src2/**',
+							'!src2/a' 
+						],
+					}
+				],
+				cssTranspilerAdapter: './fakeTranspilerAsync.js',
+			});
+
+			let error = null;
+			try {
+				await buildTranspilerFile(settings);
+			} catch (e) {
+				error = e;
+			}
+			
+			assert.equal(error, null);
+		});
 	});
 
 });