+++ test/test-buildTranspilerFile/src2/widget/widget.tpl.js
@@ -1,0 +1,6 @@
+module.exports = (tpl, props) => tpl`
+	<div class="mojl-widget">
+		<div>${props.clicks}</div>
+		<button>increment</button>
+	</div>
+`;