+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Made collation names optional
 
+### Removed
+- `buildTempDir` option
+
+### Fixed
+- temp dir deletion collision
+
 
 ## [2.0.0-alpha.8] - 2023-08-21